+++ hyperspy/tests/signals/test_lazy.py
@@ -151,17 +151,6 @@ def test_rechunk(signal, nav_chunks, sig_chunks):
                    sig_chunks=sig_chunks)
 
 
-def test_warning():
-    sig = _signal()
-
-    with pytest.warns(VisibleDeprecationWarning, match="progressbar"):
-        sig.compute(progressbar=False)
-
-    assert sig._lazy == False
-    thing = to_array(sig, chunks=None)
-    assert isinstance(thing, np.ndarray)
-
-
 class TestGetNavigationDimensionHostChunkSlice:
     @pytest.mark.parametrize(
         "position, chunk_slice",